+++ examples/actions/nxos_save_config/action.tf
@@ -1,0 +1,5 @@
+action "nxos_save_config" "example" {
+  config {
+    device = "switch1"
+  }
+}