+++ examples/resources/nxos_rest/import-by-identity.tf
@@ -1,0 +1,7 @@
+import {
+  to = nxos_rest.example
+  identity = {
+    "dn" : "<dn>"
+    "class_name" : "<class_name>"
+  }
+}