+++ examples/resources/nxos_rest/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = nxos_rest.example
+  id = "<dn>,<class_name>"
+}